Rename delete module and update imports
