+++ delete_helpers.py
